Refactory the library to make it really go routine safe
If we decided to change the Discovery type attributes or algorithm while it was
executing we could get a race condition. To avoid that we now created Setters
that use locks to safely replace the library internals. We are also using now
an interface for the Discovery type to allow easy test mocks.