Run cover report only when it's not a PR
According to the travis documentation [1] encrypted environment variables are
not available to pull requests from forks due to the security risk of exposing
such information to unknown code.

To solve this we will add a condition in the Travis script, as described in the
Travis documentation [2].

This problem was detected in #11.

[1] https://docs.travis-ci.com/user/environment-variables/#Defining-encrypted-variables-in-.travis.yml
[2] https://docs.travis-ci.com/user/pull-requests#Pull-Requests-and-Security-Restrictions